[refact] REST API URL 양식 준수 > 게시판 그룹
